Move test_podman_login_check to destructive tests
